Model2.2 with Ben's components
Creating a prefab of this object and placing it in the scene breaks the movement but not the animations.
There are more bugs such as the model moving out of the parent (capsule collider) object due to the target or animations, affecting the pivot point of the model2.2.

Apparently, weapons create issues I can't solve now, especially weapon animations.